Unify external-write simulation into the shared base (like DSv2ExternalMutationTestBase)
The classic and Connect bases previously each implemented the external-write
simulation: classic via DeltaLog/LogStore (classic-only APIs), Connect via raw
_delta_log commit JSON on the shared filesystem. The filesystem approach needs
only java.io plus the unified DataFrame API, so it compiles in the shared
scala-shared module. Move it there once and have BOTH modes use it, mirroring
Spark's DSv2ExternalMutationTestBase which shares its external mutation via the
unified catalog API.

The shared base now provides table setup (createSimpleTable/insertInitialData/
writerSql), withRefreshTable, and externalDataWrite/externalAddColumnAndWrite/
externalDropAndRecreate. The per-module bases shrink to just the arity-mismatch
assertion (classic: Spark checkError with parameters; connect: substring tolerant
because Connect wraps the error). spark/checkAnswer/withTable/withTempPath come
from each module's test mixin via the self-type. Classic base 161->42 lines,
Connect base 207->51; no external-write duplication remains.

All 4 suites pass (12 classic + 12 connect).
